+++ blocks/landing-distances/render.php
@@ -1,9 +1,7 @@
 <?php
 /** @var array $attributes @var string $content */
-$extra = isset($attributes['className']) ? ' ' . $attributes['className'] : '';
-$wrapper = get_block_wrapper_attributes([
-  'class' => 'mh-landing-section' . $extra,
-  'id'    => $attributes['anchor'] ?? 'valimatkat',
+$wrapper = \Muuttohaukat\landing_block_wrapper_attributes($attributes, ['mh-landing-section'], [
+  'id' => $attributes['anchor'] ?? 'valimatkat',
 ]);
 ?>
 <section <?php echo $wrapper; ?>><?php echo $content; ?></section>