+++ blocks/landing-final-cta/render.php
@@ -1,9 +1,7 @@
 <?php
 /** @var array $attributes @var string $content */
-$extra = isset($attributes['className']) ? ' ' . $attributes['className'] : '';
-$wrapper = get_block_wrapper_attributes([
-  'class' => 'mh-landing-section mh-landing-section--final' . $extra,
-  'id'    => $attributes['anchor'] ?? 'tarjous',
+$wrapper = \Muuttohaukat\landing_block_wrapper_attributes($attributes, ['mh-landing-section', 'mh-landing-section--final'], [
+  'id' => $attributes['anchor'] ?? 'tarjous',
 ]);
 ?>
 <section <?php echo $wrapper; ?>><?php echo $content; ?></section>